refactor(mobile): move releaseTourGate to mvvm utils
