+++ CODEOWNERS
@@ -547,6 +547,8 @@ apps/ledger-live-mobile/src/mvvm/features/WalletV4Tour/                         
 apps/ledger-live-mobile/src/mvvm/features/ProductTour/                                         @ledgerhq/engagement
 apps/ledger-live-mobile/src/mvvm/features/Q2WalletV4Tour/                                       @ledgerhq/engagement
 apps/ledger-live-mobile/src/mvvm/features/Q3WalletV4Tour/                                       @ledgerhq/engagement
+apps/ledger-live-mobile/src/mvvm/utils/releaseTourGate.ts                                        @ledgerhq/engagement
+apps/ledger-live-mobile/src/mvvm/utils/__tests__/releaseTourGate.test.ts                         @ledgerhq/engagement
 apps/ledger-live-mobile/src/mvvm/features/BackupHub/                                           @ledgerhq/engagement
 apps/ledger-live-mobile/src/mvvm/features/LNUpsell/                                           @ledgerhq/engagement
 apps/ledger-live-mobile/src/mvvm/features/LargeScreenUpsell/                                   @ledgerhq/engagement
