+++ apps/ledger-live-mobile/src/mvvm/features/Q2WalletV4Tour/Drawer/hooks/useQ2WalletV4TourDrawerViewModel.ts
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "~/context/hooks";
 import { setHasSeenQ2WalletV4Tour } from "~/actions/settings";
 import { hasSeenQ2WalletV4TourSelector } from "~/reducers/settings";
 import { track } from "~/analytics";
-import { isQ2ReleaseTourEnabled } from "../../releaseTourGate";
+import { isQ2ReleaseTourEnabled } from "LLM/utils/releaseTourGate";
 import { PAGE_TRACKING_Q2_WALLET_V4_TOUR } from "../const";
 import type { Q2WalletV4TourDrawerViewModel } from "../types";
